requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-hashid-field==3.1.3
django-import-export==2.5.0
django-jazzmin==2.4.4
django-mssql==1.8
django-mssql-backend==2.8.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
hashids==1.3.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
psycopg2==2.8.6
pyodbc==4.0.30
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
